pinctrl: bcm: ns: Use uintptr_t for data
Fix up a compiler error on 64bit architectures where pointers
and integers differ in size.

Cc: Rafał Miłecki <rafal@milecki.pl>
Suggested-by: Arnd Bergmann <arnd@arndb.de>

+++ drivers/pinctrl/bcm/pinctrl-ns.c
@@ -20,7 +20,7 @@
 
 struct ns_pinctrl {
 	struct device *dev;
-	unsigned int chipset_flag;
+	uintptr_t chipset_flag;
 	struct pinctrl_dev *pctldev;
 	void __iomem *base;
